Update to 1.1-Alpha
This will... not make any difference, there are various bugs fixed and welp, it's time to add support to other devices. Retro Music Player will be added not because we got permission, it's because we love it and sorry if that bothers to the dev. It's just our way to appreciate his hard work on that player and the honor of having it as default.