vendor: Implement our Bootleggers bootanimation
This will make a bootanimation for all the devices without the need of making a bunch of zips, also the half-res is supported in case people don't care about quality and want a bit of MBs less on their zip.

Also, it's based in LineageOS way to implement it. Thanks to them and also some fixes taken from AICP.

Just for the record, the fixes are:

vendor: fix bootanimation
https://github.com/AICP/vendor_aicp/commit/c5c0827b90ed5d7728b532152a5995c37d8f8686#diff-7c7861cd14ef1252c848b53fe5a3ef7c

Revert "cm: bootanimation: Fail the build if mogrify is absent"
https://github.com/AICP/vendor_aicp/commit/2678ca266cd09570814afd12c23a21357afc9fee#diff-7c7861cd14ef1252c848b53fe5a3ef7c

Revert "bootanimation: Use mogrify to convert images in place"
https://github.com/AICP/vendor_aicp/commit/731856cc5ff363593ec0434c345d362649c43d0f#diff-7c7861cd14ef1252c848b53fe5a3ef7c

Let's say goodbye to the classic AOSP one, it won't be missed