vendor: Fix markup building
From many months we got this massive derp and that made many people run builds without a proper image editor for their screenshots.

It was never noticed due to markup being bundled within gapps so, in devices that flash gapps that doesn't add markup, now they'll have the screenshot editor.

And also update the prebuilt to the latest release

Co-authored-by: shagbag913 <sh4gbag913@gmail.com>