vendor: New wallpaper with some extra changes
This is... something by eldainosor.

UPDATE: Let's switch between our traditional method to set a wallpaper with an overlay just for now.
Maybe research a method to crop all the images with imagemagick to adapt them on the 18:9-19:9 devices.

for now, those devices should have BOOTLEGGERS_NOTCHED set as true to get the actual wallpaper cropped out.