vendor: Switch from BOOTLEGGERS_NOTCHED to BOOTLEGGERS_IS_LARGE_DEVICE
This is for our custom overlays for larger aspect ratio devices (18:9)

We used for a long time BOOTLEGGERS_NOTCHED but it wasn't a descriptive
var export. So let's make it clearer and enabled by default because
most of devices nowadays are large lads.

Until I find out how to run a script to do the whole sizing and just throwing
a png instead of generating 6 manually